Greatly simplify the whole idea behind the template
The new approach is a popular "convention over configuration" approach
which is not ideal, but should work well enough granted all conventions
are provided as a part of this repo.

The new structure is based on my master thesis and I wanted it to be as
simple as possible. That's why there's no longer a new document class or
some over-engineered git submodule approach to just use this thing.

I'm not sure when I'll be back to using LaTeX, but I'd like to just came
back to this repo to remind myself some good LaTeX practices if this
ever happens. The whole solution should be pretty much universal,
PG-stuff is mainly related to the concrete choices in the configuration
options, but the core and the "glue" which holds this template together
will be probably the same.

At the same time, now it might be actually useful for someone out there.