Resign from the English version of the template
Make Polish version the only supported version, but explain how to make
it work in English.

At this point I'm not 100% sure that this instruction and changing
4 lines covers everything, but it was exactly like that when I decided
to support both versions. This commit is mainly made to safely get rid
of 'polish' branch and merge it to main, and I'll be replacing the whole
approach afterwards, so there's no sense in making the investigation at
this point.